+++ scripts/dzVents/runtime/dzVents.lua
@@ -24,6 +24,12 @@ if (tonumber(globalvariables['dzVents_log_level']) == utils.LOG_DEBUG) then
 end
 --
 
+-- todo remove!!!!!!
+--print('!!!!!!!!!! remove')print('!!!!!!!!!! remove')
+local persistence = require('persistence')
+persistence.store(currentPath .. '/domoticzData.lua', domoticzData)
+
+
 commandArray = {}
 
 utils.log('dzVents version: 2.2.1', utils.LOG_DEBUG)